labgrid: rpi: Rework the writer to be more robust
In order to always correctly write the correct file to the correct location, we
need to first ensure that our newly generated config.txt file is readable to
everyone, as it may be accessed by another user during transport. Second, as
this is a file for the first partition, tell write-files this rather than
assume it.

Signed-off-by: Tom Rini <trini@konsulko.com>