labgrid: sage: Add new lab
This adds support for the "sage" lab, that is utilizing stock labgrid to
control a Raspberry Pi 3, a Raspberry Pi 4, a Pine64 Plus and a SolidRun
Hummingboard 2. The board environment files enable additional tests by
pointing to content which is any of:
- Copy/pasted from the documentation for a test.
- Obvious in its source (e.g. Linux Kernel or GRUB).
- Random content (1MiBtest.bin was created with dd from /dev/urandom).

The file bin/sage/labgrid-sage-env.yaml is the file normally used either in the
LG_ENV environment variable or passed to labgrid-client as the environment to
use. This shows that each of the above targets is attached to its own labgrid
exporter node and so all parts of the system can be power cycled to help clear
transient USB errors.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Changes in v2:
- Add pine64-1 and update the commit message.
- Some updates to better match upstream best practices.