labgrid: sage: Add BeagleBone Black
This adds support for a BeagleBone Black board, which is supported under the
am335x_evm configuration.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
For the moment, this uses snmpset to manage power on the board. The relevant
changes have been submitted upstream, but rather than use a local wheel (which
is doable but complicated for all custodian repositories) we go with this option
instead for now.